Use frontend-i18n package instead of local i18n directory.
Delete the local copy of the i18n library.  Note: the package.json path MUST be updated prior to merging this PR.
