--- src/i18n/index.jsx
@@ -1,33 +1,0 @@
-import { intlShape } from 'react-intl';
-import injectIntlWithShim from './injectIntlWithShim';
-import {
-  getCountryList,
-  getCountryMessages,
-  getLanguageList,
-  getLanguageMessages,
-  getLocale,
-  getAssumedServerLanguageCode,
-  getMessages,
-  handleRtl,
-  isRtl,
-} from './i18n-loader';
-import { setLocale } from './actions';
-import reducer from './reducers';
-import { localeSelector } from './selectors';
-
-export {
-  injectIntlWithShim as injectIntl,
-  getCountryList,
-  getCountryMessages,
-  getLanguageList,
-  getLanguageMessages,
-  getLocale,
-  getAssumedServerLanguageCode,
-  getMessages,
-  handleRtl,
-  isRtl,
-  intlShape,
-  setLocale,
-  reducer,
-  localeSelector,
-};
