--- src/i18n/selectors.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export const localeSelector = state => state.i18n.locale;
