+++ src/index.jsx
@@ -10,13 +10,14 @@ import {
 } from '@edx/frontend-analytics';
 import { configureLoggingService, NewRelicLoggingService } from '@edx/frontend-logging';
 import { getAuthenticatedAPIClient } from '@edx/frontend-auth';
+import { handleRtl, configure as configureI18n } from '@edx/frontend-i18n';
 
 import { configuration } from './environment';
-import { handleRtl } from '@edx/frontend-i18n'; // eslint-disable-line
 import configureStore from './store';
 import { configureUserAccountApiService } from './common';
 import { configureApiService as configureAccountSettingsApiService } from './account-settings';
 import { configureApiService as configureSiteLanguageApiService } from './site-language';
+import messages from './i18n';
 
 import './index.scss';
 import App from './components/App';
@@ -53,6 +54,8 @@ function createInitialState() {
 }
 
 function configure() {
+  configureI18n(configuration, messages);
+
   const { store, history } = configureStore(createInitialState(), configuration.ENVIRONMENT);
 
   configureLoggingService(NewRelicLoggingService);
