+++ src/site-language/service.js
@@ -1,5 +1,3 @@
-import { getAssumedServerLanguageCode } from '@edx/frontend-i18n'; // eslint-disable-line
-
 import siteLanguageList from './constants';
 import { snakeCaseObject, convertKeyNames } from '../common/utils';
 import applyConfiguration from '../common/serviceUtils';
@@ -36,7 +34,7 @@ async function patchPreferences(username, params) {
 
 async function postSetLang(code) {
   const formData = new FormData();
-  formData.append('language', getAssumedServerLanguageCode(code));
+  formData.append('language', code);
 
   await apiClient.post(`${config.LMS_BASE_URL}/i18n/setlang/`, formData, {
     headers: { 'X-Requested-With': 'XMLHttpRequest' },
